+++ src/neuroconv/tools/data_transfers/__init__.py
@@ -1,12 +1,11 @@
 """Collection of helper functions for assessing and performing automated data transfers."""
 
-from ._aws import estimate_s3_conversion_cost, submit_aws_batch_job, estimate_total_conversion_runtime
+from ._aws import estimate_s3_conversion_cost, submit_aws_batch_job
 from ._dandi import automatic_dandi_upload
 from ._globus import get_globus_dataset_content_sizes, transfer_globus_content
 from ._helpers import estimate_total_conversion_runtime
 
 __all__ = [
-    "estimate_total_conversion_runtime",
     "submit_aws_batch_job",
     "estimate_s3_conversion_cost",
     "automatic_dandi_upload",