+++ tests/test_minimal/test_tools/s3_tools.py
@@ -1,6 +1,7 @@
 from neuroconv.tools.data_transfers import (
     estimate_s3_conversion_cost,
     estimate_total_conversion_runtime,
+    submit_aws_batch_job,
 )
 
 
@@ -46,3 +47,38 @@ def test_estimate_total_conversion_runtime():
         1235294.1176470588,
         12352941.176470589,
     ]
+
+
+def test_submit_aws_batch_job():
+    job_name = "test_submit_aws_batch_job"
+    docker_image = "ubuntu:latest"
+    command = "echo 'Testing NeuroConv AWS Batch submission."
+
+    submit_aws_batch_job(
+        job_name=job_name,
+        docker_image=docker_image,
+        command=command,
+    )
+
+
+def test_submit_aws_batch_job_with_dependencies():
+    job_name_1 = "test_submit_aws_batch_job_with_dependencies_1"
+    docker_image = "ubuntu:latest"
+    command_1 = "echo 'Testing NeuroConv AWS Batch submission."
+
+    info = submit_aws_batch_job(
+        job_name=job_name_1,
+        docker_image=docker_image,
+        command=command_1,
+    )
+    job_submission_info = info["job_submission_info"]
+
+    job_name_2 = "test_submit_aws_batch_job_with_dependencies_1"
+    command_2 = "echo 'Testing NeuroConv AWS Batch submission with dependencies."
+    job_dependencies = [{"jobId": job_submission_info["jobId"], "type": "SEQUENTIAL"}]
+    submit_aws_batch_job(
+        job_name=job_name_2,
+        docker_image=docker_image,
+        command=command_2,
+        job_dependencies=job_dependencies,
+    )